requirements.txt

hikari
hikari-tanjun
python-dotenv
pathlib2
typing
asyncio
pyfiglet
colorama
datetime
git-python
apscheduler
nox
flake8
black
isort
asyncpg
pokebase
pillow
python-dateutil
